Smarter approach to normalization
rather than fussing with what the decoding library spits out, just
normalize both.

This may fix the tests, but they take a very long time to run, and I
have to get out the door. Will check again later. Note that I tried
changing the infinite loop to a debug.todo and it was still taking
forever, so it's not that.